server/requirements.txt
flask
amrlib
textacy
benepar
stanza
spacy_stanza
unidecode
waitress
git+https://github.com/hltcoe/PredPatt.git